+++ src/eca/models.clj
@@ -46,6 +46,7 @@
                (assoc p (str provider "/" model)
                       (assoc-some
                        {:reason? (get model-config "reasoning")
+                        :image-input? (contains? (set (get-in model-config ["modalities" "input"])) "image")
                         ;; TODO how to check for web-search mode dynamically,
                         ;; maybe fixed after web-search toolcall is implemented
                         :web-search (contains? models-with-web-search-support (str provider "/" model))