+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `/compact` command to summarize the current conversation helping reduce context size.
+- Add support for images as contexts.
 
 ## 0.52.0